Version 2.1BETA
Major changes - 
*Created a new route for model3 in app.py
*created a new function for model3 in genai_controller.py
*created a new file named model3_prompt.;py

Difference between earlier model and this model - 
The major difference is that it takes another input from the user besides the images before creating a response. The input is "Number of Questions" to generate in the quiz. So I had to add the numberOfQuestions in the quiz so i went for a .py file instead of a txt file.